// src/app/models/curriculo.ts
export class Curriculos {
  bairro: string;
  categoria: string;
  celular: string;
  cidade: string;
  cpf: string;
  createdAt: string;
  dataNascimento: string;
  endereco: string;
  estado: string;
  filhos: string;
  telefone: string;
  id: number;
  idade: string;
  idadeFilhos: string;
  moradiaAntes: string;
  nome: string;
  nomeMae: string;
  nomePais: string;
  numeroCarteiraMotorista: string;
  numeroCarteiraProfissional: string;
  pis: string;
  proximo: string;
  rg: string;
  secaoEleitoral: string;
  serieCarteiraProfissional: string;
  tempoMoradia: string;
  tempocidade: string;
  ufCarteiraProfissional: string;
  updatedAt: string;
  zonaEleitoral: string;
  foneRecador: string;
  localNascimento: string;
  orgaoEmissor: string;
  dataEmissao: string;
  tituloEleitor: string;
  casaPropria: string;
  valorAluguel: string;
  rgestado: string;
}